requirements.txt
fair-esm
torch 

pandas
matplotlib
scikit-learn
jupyter

tqdm 


